Allowed for incremential builds

+++ cmds/buildtar.py
@@ -1,6 +1,10 @@
 import tarfile
 import os
 
+if not os.path.exists("src/sysroot/"):
+	print("SYSROOT DOES NOT EXIST")
+	exit(1)
+
 if os.path.exists("src/sysroot.img"):
 	os.unlink("src/sysroot.img")
